align bridge history reasoning replay
