+++ docs/chat-chain-changes/2026-06-19-bridge-history-reasoning-content.md
@@ -1,0 +1,6 @@
+---
+date: 2026-06-19
+pr: pending
+feature: Bridge history reasoning replay
+impact: Chat runs now use the Agent Bridge path only and preserve empty assistant reasoning_content sentinels when replaying history for providers such as DeepSeek and MiMo.
+---
