+++ tests/server/run-chat-message-format.test.ts
@@ -8,41 +8,12 @@ vi.mock('../../packages/server/src/services/logger', () => ({
 }))
 
 import {
-  convertHistoryFormat,
   handleMessage,
   isAssistantMessageSendable,
 } from '../../packages/server/src/services/hermes/run-chat/message-format'
 import type { SessionMessage } from '../../packages/server/src/services/hermes/run-chat/types'
 
 describe('run-chat message formatting', () => {
-  it('drops empty assistant history messages without tool calls', () => {
-    const formatted = convertHistoryFormat([
-      { role: 'user', content: 'run a command' },
-      { role: 'assistant', content: '' },
-      { role: 'user', content: 'next turn' },
-    ])
-
-    expect(formatted).toEqual([
-      { role: 'user', content: 'run a command' },
-      { role: 'user', content: 'next turn' },
-    ])
-  })
-
-  it('converts empty assistant tool-call history messages to non-empty text', () => {
-    const toolCalls = [{
-      id: 'call_1',
-      type: 'function',
-      function: { name: 'terminal', arguments: '{}' },
-    }]
-    const formatted = convertHistoryFormat([
-      { role: 'assistant', content: '', tool_calls: toolCalls },
-    ])
-
-    expect(formatted).toEqual([
-      { role: 'assistant', content: '[Calling tool: terminal with arguments: {}]' },
-    ])
-  })
-
   it('drops stale empty assistant messages loaded from the session database', () => {
     const messages: SessionMessage[] = [
       { id: 1, session_id: 's1', role: 'user', content: 'first', timestamp: 1 },
